Initial commit. See README, doc, and NOTES.
This version of Zen has been shaped to create an experimental environment
for developing continuation-based, straight-line (linear) coded programmes.
To write a simple but significant programme, we want 'system calls' and
library functions that block so we can write straight-line code the way
we call read and write in C or puts and gets in Ruby. This version of Zen
contains a probably broken version of jsScheme that can support
continuations.

This version of Zen tries to simplify future development and debugging of
Zen by removing its dependencies upon Ruby on Rails.